Don't save block hash in S3 metadata.
Summary:
Rarely, for unclear reasons, when we save a block to S3 the data_sha256
metadata doesn't get set. This breaks sync if you try to save the same
block again later, because the saved metadata doesn't match the hash.

In any case, setting this metadata is patently pointless, because the S3 object
key is already the hash. So just remove this code.

Fixes T758.

Test Plan: Stare at code.

Reviewers: kav-ya

Reviewed By: kav-ya

Maniphest Tasks: T758

Differential Revision: https://review.inboxapp.com/D1073